# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Rafiki
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /cursos:
    # our controller name
    x-swagger-router-controller: cursos
    get:
      description: Obtener la lista de cursos
      operationId: getCursos
      # define the type of response for Success "200" and Error
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseCurso"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    post:
      description: Añadir un nuevo curso
      operationId: postCurso
      # movie info to be stored
      parameters:
        - name: data
          description: Atributos de curso
          in: body
          required: true
          schema:
            $ref: "#/definitions/Curso"     
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
...
  /cursos/{id}:
    # our controller name
    x-swagger-router-controller: cursos
    get:
      description: Obtener un curso
      operationId: getCurso
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ResponseCurso"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
    put:
      description: Actualizar un curso
      operationId: updateCurso
      # define the parameters
      parameters:
        - name: id
          description: Curso id
          type: string
          in: path
          required: true
        - name: name
          description: Propiedades de User
          in: body
          required: true
          schema:
            $ref: "#/definitions/Curso"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"        
    delete:
      description: Eliminar curso
      operationId: deleteCurso
      # define the parameters
      parameters:
        - name: id
          description: User id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Response"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Errors"
...
definitions:
...
  Curso:
    type: object
    properties:
      idCurso:
        type: string
        description: Sigla del curso (Ejemplo 1A)
      nivel:
        type: string
        description: Nivel educacional del curso
      asignatura:
        type: string
        description: Nombre de la asignatura
      profesorJefe:
        type: string
        description: Nombre del profesor jefe del curso
      salaCurso:
        type: string
        description: Sala del curso
      totalAlumnos:
        type: number
        description: Cantidad total de alumnos del curso
...
  Errors:
    type: object
    properties:
      errors:
        type: array
        items:
          type: object
          properties:
            titulo:
              type: string
            detalle:
              type: string
            link:
              type: string
            estado:
              type: string
...
  Response:
    type: object
    properties:
      link:
        type: string
# Objetivos de Aprendizaje
...
  ResponseCurso:
    type: object
    properties:
      link:
        type: string
      data:
        type: array
        items:
          $ref: "#/definitions/Curso"    
      type:
        type: string   